refactor(app_framework): split middleware boundaries
